web_scrapingnew/requirements.txt
fastapi
uvicorn
google-generativeai
beautifulsoup4
requests
pydantic
python-multipart
python-dotenv